# fixes issue #79