* handles invalid \u escape sequences
* falls back to unicode replacement character "�"

```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkEJ_Unmarshal_M-8                     51707         52890         +2.29%
BenchmarkEJ_Unmarshal_S-8                     685           695           +1.46%
BenchmarkEJ_Marshal_M-8                       19262         19303         +0.21%
BenchmarkEJ_Marshal_L-8                       912472        921312        +0.97%
BenchmarkEJ_Marshal_L_ToWriter-8              830432        843928        +1.63%
BenchmarkEJ_Marshal_M_Parallel-8              5983          5789          -3.24%
BenchmarkEJ_Marshal_M_ToWriter-8              17539         17662         +0.70%
BenchmarkEJ_Marshal_M_ToWriter_Parallel-8     4249          4204          -1.06%
BenchmarkEJ_Marshal_L_Parallel-8              319229        330672        +3.58%
BenchmarkEJ_Marshal_L_ToWriter_Parallel-8     196869        200514        +1.85%
BenchmarkEJ_Marshal_S-8                       287           286           -0.35%
BenchmarkEJ_Marshal_S_Parallel-8              97.5          89.3          -8.41%

benchmark                                     old MB/s     new MB/s     speedup
BenchmarkEJ_Unmarshal_M-8                     251.88       246.24       0.98x
BenchmarkEJ_Unmarshal_S-8                     119.57       117.88       0.99x
BenchmarkEJ_Marshal_M-8                       461.61       460.63       1.00x
BenchmarkEJ_Marshal_L-8                       487.31       482.64       0.99x
BenchmarkEJ_Marshal_L_ToWriter-8              535.46       526.89       0.98x
BenchmarkEJ_Marshal_M_Parallel-8              2176.69      2249.64      1.03x
BenchmarkEJ_Marshal_M_ToWriter-8              506.98       503.45       0.99x
BenchmarkEJ_Marshal_M_ToWriter_Parallel-8     2092.38      2115.05      1.01x
BenchmarkEJ_Marshal_L_Parallel-8              1392.91      1344.71      0.97x
BenchmarkEJ_Marshal_L_ToWriter_Parallel-8     2258.66      2217.59      0.98x
BenchmarkEJ_Marshal_S-8                       282.21       283.04       1.00x
BenchmarkEJ_Marshal_S_Parallel-8              830.73       906.78       1.09x

benchmark                                     old allocs     new allocs     delta
BenchmarkEJ_Unmarshal_M-8                     128            128            +0.00%
BenchmarkEJ_Unmarshal_S-8                     3              3              +0.00%
BenchmarkEJ_Marshal_M-8                       10             10             +0.00%
BenchmarkEJ_Marshal_L-8                       29             29             +0.00%
BenchmarkEJ_Marshal_L_ToWriter-8              24             24             +0.00%
BenchmarkEJ_Marshal_M_Parallel-8              9              9              +0.00%
BenchmarkEJ_Marshal_M_ToWriter-8              8              8              +0.00%
BenchmarkEJ_Marshal_M_ToWriter_Parallel-8     8              8              +0.00%
BenchmarkEJ_Marshal_L_Parallel-8              34             34             +0.00%
BenchmarkEJ_Marshal_L_ToWriter_Parallel-8     24             24             +0.00%
BenchmarkEJ_Marshal_S-8                       1              1              +0.00%
BenchmarkEJ_Marshal_S_Parallel-8              1              1              +0.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkEJ_Unmarshal_M-8                     9792          9792          +0.00%
BenchmarkEJ_Unmarshal_S-8                     128           128           +0.00%
BenchmarkEJ_Marshal_M-8                       10377         10378         +0.01%
BenchmarkEJ_Marshal_L-8                       462278        462542        +0.06%
BenchmarkEJ_Marshal_L_ToWriter-8              2641          2641          +0.00%
BenchmarkEJ_Marshal_M_Parallel-8              10479         10477         -0.02%
BenchmarkEJ_Marshal_M_ToWriter-8              741           741           +0.00%
BenchmarkEJ_Marshal_M_ToWriter_Parallel-8     757           758           +0.13%
BenchmarkEJ_Marshal_L_Parallel-8              525769        528920        +0.60%
BenchmarkEJ_Marshal_L_ToWriter_Parallel-8     5270          4312          -18.18%
BenchmarkEJ_Marshal_S-8                       128           128           +0.00%
BenchmarkEJ_Marshal_S_Parallel-8              128           128           +0.00%
```