@cespare 2016-12-15
We've also run into this issue recently. The fix (and the code) are basically the same as encoding/json, so :+1: from me.